+++ Makefile
@@ -1,6 +1,6 @@
 CC = gcc
 CFLAGS = -Wall -Wextra -g -Wno-unused-parameter 
-SRCS = main.c my_trace.c
+SRCS = main.c
 EXEC = tracer
 INC_DIR = .
 CK_DIR = ./ck
